Now the map allows background images.
To prevent from clearing the background image after a walker pass on it, each cell now can have more than one element at the same time. Which element is visible will depend on which element was set in the coordinate last. Also, to allow a more readable map editing, it will now be split in several file layers. Each file may have a set of elements, for example, the first layer (00) may have walls and blocking elements that cannot be moved or walked on, the second layer (01) may have the background elements that can be stepped on, like grass, sand, etc., and the third (02) layer will have the autonomous walkers (NPCs). Actually, there can be any map layers and element types can be mixed among these maps, but to simplify, I suggest separating maps by the type od elements it represents. Any coordinate filled with "###" or "   " will now be ignored, walls and empty spaces must hold its own identification. This was done to be easier to identify boundaries between layers. In case of any question, check files `resources/data/level01[0?].map` (where ? is the map layer).

There are still some bugs regarding rendering again the layer when a walker step out of it, but it may be fixed in the future. In the future, the visibility priority will be set according to the layer the element is placed. Player characters should have set the highest priority. This way there will be possible to overlay elements in the map.